context/requirements_current_freeze.txt
absl-py==2.4.0
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.8.3
aiohttp_socks==0.11.0
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
aiosqlite==0.22.1
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
APScheduler==3.11.2
asyncua==1.1.8
attrs==25.4.0
audioread==3.1.0
av==14.2.0
backoff==2.2.1
beautifulsoup4==4.14.3
bitsandbytes==0.47.0
blake3==1.0.8
blend_modes==2.2.0
blind-watermark==0.4.4
boto3==1.41.5
botocore==1.41.5
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
clip-interrogator==0.6.0
cmake==4.3.1
color-matcher==0.6.0
coloredlogs==15.0.1
colour-science==0.4.7
comfy-aimdo==0.2.12
comfy-kitchen==0.2.8
comfyui-embedded-docs==0.4.3
comfyui-workflow-templates-core==0.3.192
comfyui-workflow-templates-media-api==0.3.68
comfyui-workflow-templates-media-image==0.3.117
comfyui-workflow-templates-media-other==0.3.164
comfyui-workflow-templates-media-video==0.3.71
comfyui_frontend_package==1.41.21
comfyui_workflow_templates==0.9.43
contourpy==1.3.3
cryptography==46.0.4
cstr @ git+https://github.com/WASasquatch/cstr@0520c29a18a7a869a6e5983861d6f7a4c86f8e9b
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
ddt==1.7.2
decorator==4.4.2
decord==0.6.0
deep-translator==1.11.4
defusedxml==0.7.1
diffusers==0.36.0
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
docutils==0.22.4
dotenv==0.9.9
dynamicprompts==0.30.2
e3nn==0.5.9
easydict==1.13
easyocr==1.7.2
edge-tts==7.2.7
einops==0.8.2
fairscale==0.4.13
fal_client==0.13.2
fastapi==0.115.14
ffmpy @ git+https://github.com/WASasquatch/ffmpy.git@f000737698b387ffaeab7cd871b0e9185811230d
filelock==3.17.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.1
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
glitch_this==1.0.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.70.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
googletrans-py==4.0.0
greenlet==3.3.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.16.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
img2texture @ git+https://github.com/WASasquatch/img2texture.git@d6159abea44a0b2cf77454d3d46962c8b21eb9d3
importlib_metadata==8.7.1
inference-cli==1.2.0
inference-gpu==1.2.0
inference-models==0.24.2
iniconfig==2.3.0
iopath==0.1.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
langdetect==1.0.9
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matrix-nio==0.25.2
mdurl==0.1.2
mediapipe==0.10.32
mido==1.3.3
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
mss==10.1.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
ninja==1.13.0
num2words==0.5.14
numba==0.64.0
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.0
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.24.1
onnxruntime-gpu==1.21.1
onvif-zeep-async==2.0.0
open_clip_torch==3.2.0
openai==1.109.1
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.7.0.72
openunmix==1.3.0
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.3.3
pdf2image==1.17.0
peft==0.18.1
piexif==1.1.3
pilgram==2.0.0
pillow==12.2.0
pixeloe==0.1.4
platformdirs==4.9.2
pluggy==1.6.0
plyfile==1.1.3
polars==1.39.3
polars-runtime-32==1.39.3
polygraphy==0.49.26
pooch==1.9.0
portalocker==3.2.0
proglog==0.1.12
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psd-tools==1.14.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.0.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pycryptodome==3.23.0
pycuda==2025.1.3
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
pyDeprecate==0.5.0
pydot==2.0.0
pyee==13.0.1
pygame==2.6.1
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
pylibsrtp==1.0.0
pylogix==1.0.5
pyloudnorm==0.2.0
PyMatting==1.1.15
pymodbus==3.8.3
pymunk==6.11.1
PyMuPDF==1.27.2.2
PyNaCl==1.6.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyOpenSSL==26.0.0
pyparsing==3.3.2
PyPDF2==3.0.1
pypdfium2==5.6.0
PySocks==1.7.1
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-doctr==1.0.1
python-dotenv==1.0.1
python-socks==2.8.0
pytools==2025.2.5
pytz==2026.1.post1
pyvips==2.2.3
PyWavelets==1.9.0
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.0
RapidFuzz==3.14.3
redis==5.0.8
referencing==0.37.0
regex==2026.1.15
rembg==2.0.72
reportlab==4.4.10
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rf-clip==1.1
RF-SAM-2==1.0.3
rf-segment-anything==1.0
rf_groundingdino==0.3.0
rich==14.3.2
rotary-embedding-torch==0.8.9
rpds-py==0.30.0
runwayml==4.10.0
s3transfer==0.15.0
safetensors==0.7.0
sageattention==1.0.6
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
segment-anything==1.0
segmentation_models_pytorch==0.5.0
sentencepiece==0.2.1
sentry-sdk==2.57.0
setuptools==82.0.0
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
simpleeval==1.0.6
simsimd==6.5.13
siphash24==1.8
six==1.17.0
slack_sdk==3.33.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
sox==1.5.0
soxr==1.0.0
spandrel==0.4.1
SQLAlchemy==2.0.46
starlette==0.46.2
stringzilla==4.6.0
structlog==24.4.0
supervision==0.27.0.post2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tensorrt==10.15.1.29
tensorrt_cu13==10.15.1.29
tensorrt_cu13_bindings==10.15.1.29
tensorrt_cu13_libs==10.15.1.29
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.2.24
tiktoken==0.12.0
timm==1.0.26
tldextract==5.1.3
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchcodec==0.9.0
torchscale==0.3.0
torchsde==0.2.6
torchvision==0.25.0
tqdm==4.67.3
trackers==2.3.0
trampoline==0.1.2
transformers==5.2.0
transparent-background==1.3.4
trimesh==4.11.2
triton==3.6.0
twilio==9.3.8
typer==0.16.0
typer-config==1.5.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ultralytics==8.3.40
ultralytics-thop==2.0.18
unpaddedbase64==2.1.0
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.0
validators==0.35.0
Wand==0.7.0
wandb==0.25.1
wcwidth==0.5.3
websockets==16.0
wget==3.2
wheel==0.45.1
yacs==0.1.8
yapf==0.43.0
yarl==1.22.0
zeep==4.3.2
zipp==3.23.0
zxing-cpp==2.2.0
